+++ Changelog.rst
@@ -1,6 +1,15 @@
 Change Log
 ============
 
+1.12.2
+++++++
+
+Fixes
+------
+
+* `ParallelSSHClient.copy_file` with recurse enabled and absolute destination path would create empty directory in home directory of user - #197.
+* `ParallelSSHClient.copy_file` and `scp_recv` with recurse enabled would not create remote directories when copying empty local directories.
+
 1.12.1
 ++++++